Create Seven-Display Device Simulator
Scenario
A seven-segment display.

It's a device (sometimes electronic, sometimes mechanical) designed to present one decimal digit using a subset of seven segments. 

The task was to write a program which is able to simulate the work of a seven-display device, but using single LEDs instead of segments.

Each digit is constructed from 13 LEDs (some lit, some dark, of course) - that's how we imagine it.

  #      ###      ###       #    #       ###         
  #          #           #       #    #       #          
  #      ###      ###        ###        ###        
  #      #               #             #            #
  #      ###       ###            #        ### 


The code had to display any non-negative integer number entered by the user.